% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Plot a heatmap}
\usage{
plot_heatmap(dat = NULL)
}
\arguments{
\item{dat}{Input dataframe for the heatmap.}
}
\value{
A heatmap plot object.
}
\description{
Plot a heatmap
}
